Do not store object in stash
This is an attempt to slow down the memory increase. Summary of
changes:

* The biggest culprit, REST controller, no longer stashes the objects.
* ClassicCompatibility controller no longer stashes objects either.
  Classic view code purged as well (see diff).
* Root controller appears to not make use of stashed objects.

IF ANYTHING BREAKS INEXPLICABLY, THIS IS PROBABLY THE PLACE TO
LOOK. Perhaps there are some components that do make use of stashed
objects (e.g. the template). They shouldn't be using stashed objects
anyway. That either will require a rewrite, or if the system is too
entrenched in stashed objects, this commit will have to be backed out.