Fix Datatable warning reported by Todd * added unique key DataTables warning (table id = 'table__by'): Cannot reinitialise DataTable (strains)?
To retrieve the DataTables object for this table, please pass either no arguments to the dataTable() function, or set bRetrieve to true. Alternatively, to destory the old table and create a new one, set bDestroy to true (note that a lot of changes to the configuration can be made through the API which is usually much faster).