Standardised the format of the file as it was creeping as curators interacted with the file and also corrected the "symbolic_name" syntax inline with the correct ?Database objects.
Lots of external links do not work on the site because of discrepancies between the name of thre Database object in acedb and the symbolic_name used as the primary key for fetching data in this file. One could argue that lots of the Database objects should be reviewed in acedb to be more consistent in naming but to get the links working, it has been necessary to transfer the names across to this file. There are still quite a few ?Database objects that are missing from this file:

Here are a few:
Database : "MODEncodeTissueEMB_bw_v2"
Database : "MODEncodeTissueEMB_coelomocytes"
Database : "MODEncodeTissueEMB_dopa"
Database : "MODEncodeTissueEMB_GABA"
Database : "MODEncodeTissueEMB_GLP"
Database : "MODEncodeTissueEMB_hypodermal"
Database : "MODEncodeTissueEMB_intestine"
Database : "MODEncodeTissueEMB_panneural"
Database : "MODEncodeTissueEMB_PhM"
Database : "MODEncodeTissueEMB_PVC"

So for all external links to work there is still a lot of changes to be made.